AC3: Famine Mod (CE Table)
What this script does is double up the consumed value, which is usually =1, but when you throw money, you throw 10 units of it, so with Famine you'll throw away 20. I also included something to check if the value would be below 0, because on that case it would make you have negative or infinite money/items (for example, if you have 1 bullet left, and reload, it would be reduced to -1). On that case, Famine is not applied so that issue doesn't happen (at least that's what I noticed on the three items I actually have, including money).

http://forum.cheatengine.org/viewtopic.php?t=601922